api: make the NCO-disable path of SetNCOIndex reachable
SetNCOIndex took the index as uint8_t, so the index < 0 branches that
bypass the CMIX could never be taken - the disable sentinel implied by
the int-returning GetNCOIndex (and used by the legacy LMS_SetNCOIndex)
silently wrapped to 255 and was rejected. Take a signed index, validate
the range before any register write, and skip the table selection when
disabling.

Fixes #295

+++ src/boards/LMS7002M_SDRDevice.h
@@ -40,7 +40,7 @@ class LMS7002M_SDRDevice : public SDRDevice
         uint8_t moduleIndex, TRXDir trx, uint8_t channel, uint8_t index, double frequency, double phaseOffset = -1.0) override;
 
     int GetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel) override;
-    OpStatus SetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel, uint8_t index, bool downconv) override;
+    OpStatus SetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel, int16_t index, bool downconv) override;
 
     double GetNCOOffset(uint8_t moduleIndex, TRXDir trx, uint8_t channel) override;
